@@ -1,6 +1,6 @@
 module github.com/nextdotid/proof-server
 
-go 1.17
+go 1.18
 
 require (
 	github.com/aws/aws-sdk-go-v2 v1.11.2
